Added support for gzipped files. Other small improvements.
The old version didn't support gzipped files like tar.gz and gz.tar
The script checks if the file is gzipped using `7z l $source`
If the file is gzipped then it will be extracted twice which involves some temp files and renaming.

Another improvement is the MSBuild.exe script. This version will find it on your system instead of needing the file in your PATH (which never worked for me)
Also adds a -BuildOnly switch that skips all the downloads and only builds the launcher.

And finally there's some small improvements to the code style. There aren't any real code styles online so I just made it more consistent.
We should try to find a consistent code style in the future.