Revert "Set CMDER_START to homeprofile"
This reverts commit 728e83a85bea6ad7b4ddda0b960c443e3ef9f737.

The problem with *always* setting CMDER_START is that this makes the
`-new_console:d:%USERPROFILE%` in the conemu task definitions unnecessary, as
this is now always overwritten as CMDER_START is set. This also means that a
very visible conemu UI for setting the startup dir does not work anymore which
might leave the user puzzled why.