Specify "-ib" options in "hg id" call
"hg id" by itself is not specific enough and appears to only work when
the current revision is not tagged and is not on the default branch.
When the current revision is on the default branch and isn't the tip,
there is a "attempt to concatenate field '?' (a nil value)" error.
If the revision was the tip, then the branch would appear as "tip"
rather than "(default)".

Also, remove hg related functions which are no longer used.